+++ EngineSIU/EngineSIU/Engine/Source/Editor/PropertyEditor/SkeletalMeshViewerPanel.cpp
@@ -128,8 +128,8 @@ void SkeletalMeshViewerPanel::OnResize(HWND hWnd)
 {
     RECT ClientRect;
     GetClientRect(hWnd, &ClientRect);
-    Width = ClientRect.right - ClientRect.left;
-    Height = ClientRect.bottom - ClientRect.top;
+    Width = static_cast<float>(ClientRect.right - ClientRect.left);
+    Height = static_cast<float>(ClientRect.bottom - ClientRect.top);
 }
 
 void SkeletalMeshViewerPanel::SetSkeletalMesh(USkeletalMesh* SMesh)
@@ -283,8 +283,8 @@ void SkeletalMeshViewerPanel::RenderAnimationSequence(const FReferenceSkeleton& 
     }
     UAnimDataModel* DataModel = AnimSeq->GetDataModel();
     
-    ImVec2 windowSize = ImVec2(Width*0.7, Height*0.3);
-    ImVec2 windowPos = ImVec2(0.0f, Height - windowSize.y - 30);
+    ImVec2 windowSize = ImVec2(Width*0.7f, Height*0.3f);
+    ImVec2 windowPos = ImVec2(0.0f, Height - windowSize.y - 30.f);
     ImGui::SetNextWindowSize(windowSize, ImGuiCond_Always);
     ImGui::SetNextWindowPos(windowPos, ImGuiCond_Always);
     
@@ -508,7 +508,7 @@ void SkeletalMeshViewerPanel::RenderAnimationSequence(const FReferenceSkeleton& 
                                 float NewTime = static_cast<float>(Frame) / FrameRate;
                                 if(Frame<LoopStart)
                                 {
-                                    NewTime = LoopStart / FrameRate;
+                                    NewTime = static_cast<float>(LoopStart) / FrameRate;
                                 }
                                 else if(Frame + DurationFrame >LoopEnd)
                                 {
